sisimai/string package has been renamed to moji, it stands for character or string in Japanese
