+++ moji/any.go
@@ -1,13 +1,13 @@
 // Copyright (C) 2024-2025 azumakuniyuki and sisimai development team, All rights reserved.
 // This software is distributed under The BSD 2-Clause License.
-//      _        _             
-//  ___| |_ _ __(_)_ __   __ _ 
-// / __| __| '__| | '_ \ / _` |
-// \__ \ |_| |  | | | | | (_| |
-// |___/\__|_|  |_|_| |_|\__, |
-//                       |___/ 
+//                  _ _ 
+//  _ __ ___   ___ (_|_)
+// | '_ ` _ \ / _ \| | |
+// | | | | | | (_) | | |
+// |_| |_| |_|\___// |_|
+//               |__/   
 
-package string
+package moji
 import "strings"
 
 // ContainsAny() checks whether any element in argv2 is included in argv1 or not
